requirements.txt
streamlit==1.31.0
opencv-python==4.7.0.72
torch==2.7.0
torchvision==0.22.0
numpy==1.24.3
Pillow==10.0.0
streamlit-webrtc==0.47.1
matplotlib==3.7.2
ultralytics==8.0.147 